+++ pycvvdp/video_writer.py
@@ -55,7 +55,7 @@ class VideoWriter:
             else:
                 self.process = (ffmpeg
                         .input('pipe:', format='rawvideo', pix_fmt='rgb24', s='{}x{}'.format(W, H), r=self.fps)
-                        .output(self.fname, pix_fmt='yuv420p', crf=10 )
+                        .output(self.fname, pix_fmt='yuv420p', **{ "c:v": "mpeg4", "qscale:v": "3" } ) #format='mp4', crf=10                        
                         .overwrite_output()
                         .global_args( '-hide_banner')
                         .global_args( '-loglevel', 'info' if self.verbose else 'quiet')